re-design add hospital screen
